+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
-## 0.8.2 2023-08
+## 0.8.4 2023-09
 
+## 0.8.3 2023-09-05
+
+* [Plugin API add openWindow and command.globalCallback](https://github.com/siyuan-note/siyuan/issues/9032)
+
+## 0.8.2 2023-08-29
+
+* [Improve IModel type definitions](https://github.com/siyuan-note/petal/pull/14)
+* [Add plugin type `TProtyleAction`](https://github.com/siyuan-note/siyuan/issues/8988)
 * [Add plugin event bus `destroy-protyle`](https://github.com/siyuan-note/siyuan/issues/9033)
 * [Add plugin event bus `loaded-protyle-dynamic`](https://github.com/siyuan-note/siyuan/issues/9021)